Fix APIs not being able to communicate with themselves
by binding the containers' localhost to the host IP,
so `localhost:{PORT}` will be the same in the host and in the containers.

Simple hack that took hours to be found.

In some cases, like when the resource-server tries to fetch OIDC metadata from keycloak
(http://localhost:{KEYCLOAK_PORT}/realms/{REALM_NAME}/.well-known/openid-configuration)
before validating the first token, the `localhost` address wasn't being resolved
to the hosts'. This is by design, for reliability and security reasons, but it
works sometimes (like when the client-backend asks tokens to keycloak) which
makes things confusing.

Docker compose creates a network for the services where each name are resolved
to the respective container IP (See https://docs.docker.com/compose/networking/).

But using the service names instead of `localhost` didn't work out well, because:
 - The host's browser doesn't have these DNS resolutions, the app is still
   accessed from `http://localhost:{CLIENT_PORT}`
 - Without configuration, Keycloak issues tokens and bases the authorization
   endpoints according to the incoming requests' URL
   (`http://localhost:{KEYCLOAK_PORT}/`)
   (See `https://www.keycloak.org/server/hostname`)
 - Tokens couldn't be validated due to the issuer
   `http://localhost:{KEYCLOAK_PORT}/` being different from the URLs set in the
   APIs (`http://{SERVICE_NAME}:{KEYCLOAK_PORT}`)
 - Configuring different Keycloak endpoints introduces several problems, like
   the default themes not working anymore
   (See https://keycloak.discourse.group/t/stuck-understanding-frontendurl-baseurl-and-configuring-registration-pages/4638)

Another solution would be to make all services run in the `host` network
(See https://docs.docker.com/network/host/), but this only works in Docker for
Linux, and even if using WSL, some unstable workarounds are required to reach
the network from Windows (where the browser process is running).