Add "View token" button in the client app
The public key sent as parameter to jwt.io will probably be invalid
if the default keycloak port or realm name are chaged, but
that's just a minor incovenience. That can be fixed by fetching the
certs url from the "/.well-known/openid-configuration" endpoint
and getting the public key from there.